replace...?

+++ examples/countup.go
@@ -3,8 +3,8 @@ require (
 )
 
 func main() noreturn {
-	declare res int; 
-	countup x int from 0 to 10 { 
+	declare res int64;
+	countup x int64 from 0 to 10 {
 		res = x; 
 	};
 	std::os::exit_with(res);
