+++ examples/declare_autovar.go
@@ -3,7 +3,7 @@ require (
 )
 
 func main() noreturn {
-	declare x int;
+	declare x int64;
 	x = 30;
 	std::os::exit_with(x);
 }
