+++ examples/declare_twovar.go
@@ -2,8 +2,8 @@ require (
     "std"
 )
 func main() noreturn {
-	declare x int;
-	declare y int;
+	declare x int64;
+	declare y int64;
 	x = y = 3;
 	std::os::exit_with(x * y);
 }
