+++ examples/without_argument.go
@@ -1,7 +1,7 @@
 require (
     "std"
 )
-func foo() int {
+func foo() int64 {
 	return 30; 
 }
 func main() noreturn {
